improved views & rest api

+++ src/main/java/com/elderbyte/vidada/service/MediaLibraryService.java
@@ -109,7 +109,7 @@ public class MediaLibraryService  {
 
     @Transactional
 	public void addLibrary(final MediaLibrary lib){
-        repository.save(lib);
+        save(lib);
         onLibraryAdded(lib);
         libraryAddedEvent.fireEvent(this, EventArgsG.build(lib));
 	}
@@ -118,31 +118,35 @@ public class MediaLibraryService  {
 	public void removeLibrary(final MediaLibrary library) {
         MediaLibrary libToDelete = repository.findOne(library.getId());
 
+        if(libToDelete != null){
+            // Find all medias which have this library
+            List<MediaItem> mediaItems = mediaRepository.queryByLibrary(library);
 
-        // Find all medias which have this library
-        List<MediaItem> mediaItems = mediaRepository.queryByLibrary(library);
+            for (MediaItem mediaItem : mediaItems) {
+                for (MediaSource source : new HashSet<>(mediaItem.getSources())) {
+                    if(source.getParentLibrary().equals(library)) {
+                        mediaItem.getSources().remove(source);
+                    }
+                }
 
-        for (MediaItem mediaItem : mediaItems) {
-            for (MediaSource source : new HashSet<>(mediaItem.getSources())) {
-                if(source.getParentLibrary().equals(library)) {
-                    mediaItem.getSources().remove(source);
+                // Check if the media has any sources left
+                if(mediaItem.getSources().isEmpty()){
+                    // The media has no sources left
+                    // TODO We can delete it now, but then we loose all general info - is this intended?
+                    mediaRepository.delete(mediaItem);
                 }
             }
 
-            // Check if the media has any sources left
-            if(mediaItem.getSources().isEmpty()){
-                // The media has no sources left
-                // TODO We can delete it now, but then we loose all general info - is this intended?
-                mediaRepository.delete(mediaItem);
-            }
+            repository.delete(libToDelete);
+            libraryRemovedEvent.fireEvent(this, EventArgsG.build(library));
+        }else{
+            log.warn("Could not delete library " + library + " since it was not found in the database!");
         }
-
-        repository.delete(libToDelete);
-        libraryRemovedEvent.fireEvent(this, EventArgsG.build(library));
 	}
 
     @Transactional
 	public void save(final MediaLibrary lib) {
+        log.info("Saving media library " + lib);
         repository.save(lib);
 	}
